change storage's controller to interface.
