+++ src/main/java/org/signal/storageservice/controllers/StorageController.java
@@ -30,7 +30,7 @@ import javax.ws.rs.core.Response.Status;
 import org.signal.storageservice.auth.User;
 import org.signal.storageservice.metrics.UserAgentTagUtil;
 import org.signal.storageservice.providers.ProtocolBufferMediaType;
-import org.signal.storageservice.storage.StorageItemsTable;
+import org.signal.storageservice.storage.bigtable.StorageItemsTable;
 import org.signal.storageservice.storage.StorageManager;
 import org.signal.storageservice.storage.protos.contacts.ReadOperation;
 import org.signal.storageservice.storage.protos.contacts.StorageItems;
