+++ src/main/java/org/signal/storageservice/storage/bigtable/GroupLogTable.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-package org.signal.storageservice.storage;
+package org.signal.storageservice.storage.bigtable;
 
 import static com.codahale.metrics.MetricRegistry.name;
